# src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Students Compass
  description: |
    <h3>
    Description:
    <h5>Great API for students looking for a place to stay and spend your spare time on.
    <h3>
    Authors:
    <ul>
    <li>Ávila Maqueda, María del Mar
    <ul> - Correo: maravimaq@alum.us.es</ul>
    <li>González Marcos, Pedro
    <ul> - Correo: pedgonmar2@alum.us.es</ul>
    <li>Ruiz Delgado, Victoria del Carmen
    <ul> - Correo: vicruidel1@alum.us.es </ul>
    <li>Tomás Vela, Elena
    <ul> - Correo: eletomvel@alum.us.es</ul>
    <li>Zarzuela Reina, Carlos
    <ul> - Correo: carzarrei@alum.us.es</ul>
  termsOfService: http://swagger.io/terms/
  contact:
    email: studentscompass007@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://student-compass-350106.ew.r.appspot.com/api
  description: AppEngine server
- url: "http://localhost:{port}/api"
  description: Development server
  variables:
    port:
      enum:
      - "8089"
      default: "8089"
...
paths:
...
  /places/{placeId}:
    get:
      tags:
      - places
      summary: Get a place by id
      description: Get a place by id
      operationId: getAPlace
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
        examples:
          Estanislao Id:
            value: 5
          ETSII Id:
            value: 0
      responses:
        "200":
          description: Values retrieved of a particular place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
              examples:
                Catedral de Sevilla:
                  $ref: '#/components/examples/CatedralSevilla'
                Test Place:
                  $ref: '#/components/examples/TestPlace'
                Estanislao:
                  $ref: '#/components/examples/Estanislao'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - places
      summary: Update a place by id
      description: Update place information given the place id.
      operationId: updatePlace
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 4
      requestBody:
        description: Response body of the data that wants to be updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Place'
            examples:
              Update Test:
                $ref: '#/components/examples/UpdateTest'
              Test:
                $ref: '#/components/examples/TestPlace'
      responses:
        "200":
          description: Place was updated sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
              examples:
                Place1:
                  $ref: '#/components/examples/CatedralSevilla'
                Place2:
                  $ref: '#/components/examples/TestPlace'
                Estanislao:
                  $ref: '#/components/examples/Estanislao'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - places
      summary: Delete a place
      description: Deletes the place with the given id.
      operationId: deletePlace
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 4
      responses:
        "204":
          description: The resource was succesfully deleted.
        "404":
          $ref: '#/components/responses/NotFound'
  /places/{placeId}/reviews:
    get:
      tags:
      - reviews
      summary: Get all reviews from a place by id
      description: Get all reviews from a place by id
      operationId: getAllReviews
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: List of all the reviews from a particular place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviews'
        "404":
          $ref: '#/components/responses/NotFound'
    post:
      tags:
      - reviews
      summary: Creates new reviews
      description: Creates new reviews with the given payload to the place.
      operationId: addReviews
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      requestBody:
        description: Payload of the review
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        "201":
          description: Succesfully created place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "400":
          $ref: '#/components/responses/BadRequest'
  /places/{placeId}/reviews/{reviewId}:
    get:
      tags:
      - reviews
      summary: Get the specified review
      description: Get the specified review from a place by id.
      operationId: getReview
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      - name: reviewId
        in: path
        description: Id of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: Specified review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
              examples:
                Review 1:
                 $ref: '#/components/examples/review1'
                Review 5:
                 $ref: '#/components/examples/review5'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - reviews
      summary: Update the specified review
      description: Update the specified review.
      operationId: updateReview
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      - name: reviewId
        in: path
        description: Id of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      requestBody:
        description: Payload of the updated review
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        "200":
          description: Review successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
              examples:
                Review 3:
                 $ref: '#/components/examples/review3'
                Review 2:
                 $ref: '#/components/examples/review2'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - reviews
      summary: Deletes a review
      description: Deletes the review with the given id
      operationId: deleteReview
      parameters:
      - name: placeId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      - name: reviewId
        in: path
        description: Id of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "204":
          description: Review successfully deleted
        "404":
          $ref: '#/components/responses/NotFound'
...
  /events:
    get:
      tags:
      - events
      summary: Get all events
      description: Get all events
      operationId: getAllEvents
      parameters:
        - name: initialDate
          in: query
          description: This field sets the initial date for events to be returned. You can specify it by giving one string whith the following format yyyy-MM-dd. Anyway, it doesn't need to be filled. If you don´t fill the initial date it will show you all the events since the current date.
          required: false
          schema:
            type: string
            format: date
            pattern: "^\\d\\d\\d\\d-(0?[1-9]|1[0-2])-(0?[1-9]|[12][0-9]|3[01])$"
            example: "2022-08-10"
        - name: finalDate
          in: query
          description: This field sets final date for events to be returned. You can specify it by giving one string whith the following format yyyy-MM-dd. Anyway, it doesn't need to be filled. If you don´t fill the final date it shows you all the events that are hold for a month since the actual date.
          required: false
          schema:
            type: string
            format: date
            pattern: "^\\d\\d\\d\\d-(0?[1-9]|1[0-2])-(0?[1-9]|[12][0-9]|3[01])$"
            example: "2022-09-10"
        - name: order
          in: query
          description: This parameter sort the events by date ascending (+date) or descending order (-date).
          required: false
          schema:
            type: string
            example: "+date"
      responses:
        "200":
          description: Returns a list of events
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Events'
        "400":
          $ref: '#/components/responses/BadRequest'
    post:
      tags:
      - events
      summary: Creates new events
      description: Creates new events with the given payload
      operationId: addEvents
      requestBody:
        description: Payload of the event
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
            examples:
              Add event:
                $ref: '#/components/examples/Event1'
              Add another event:
                $ref: '#/components/examples/Event2'
        required: true
      responses:
        "201":
          description: Succesfully created event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                Event1:
                  $ref: '#/components/examples/Event1'
                Event2:
                  $ref: '#/components/examples/Event2'
        "400":
          $ref: '#/components/responses/BadRequest'
  /events/{eventId}:
    get:
      tags:
      - events
      summary: Get a event by id
      description: Get a event by id
      operationId: getEvent
      parameters:
      - name: eventId
        in: path
        description: Id of the event
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: Values retrieved of a particular event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                Event1:
                  $ref: '#/components/examples/Event1'
                Event2:
                  $ref: '#/components/examples/Event2'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - events
      summary: Update the specified event
      description: Update the specified event
      operationId: updateEvent
      parameters:
      - name: eventId
        in: path
        description: Id of the event
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      requestBody:
        description: Payload of the updated event
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
            examples:
              Event Update Test:
                $ref: '#/components/examples/EventTest'
                
      responses:
        "200":
          description: Event was succesfully updated
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - events
      summary: Deletes an event
      description: Deletes the event with the given id
      operationId: deleteEvent
      parameters:
      - name: eventId
        in: path
        description: Id of the event
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "204":
          description: Event successfully deleted
        "404":
          $ref: '#/components/responses/NotFound'
  /events/{eventId}/reviews:
    get:
      tags:
      - reviews
      summary: Get all reviews from a event by id
      description: Get all reviews from a event by id
      operationId: getAllReviewsEvents
      parameters:
      - name: eventId
        in: path
        description: Id of the event
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: List of all the reviews from a particular event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviews'
        "404":
          $ref: '#/components/responses/NotFound'
    post:
      tags:
      - reviews
      summary: Creates new reviews
      description: Creates new reviews with the given payload to the event
      operationId: addReviewsEvent
      parameters:
      - name: eventId
        in: path
        description: Id of the event
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      requestBody:
        description: Payload of the review
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        "201":
          description: Succesfully created event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        "400":
          $ref: '#/components/responses/BadRequest' 
  /events/{eventId}/reviews/{reviewId}:
    get:
      tags:
      - reviews
      summary: Get the specified review
      description: Get the specified review from a event by id
      operationId: getReviewEvent
      parameters:
      - name: eventId
        in: path
        description: Id of the event
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      - name: reviewId
        in: path
        description: Id of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: Specified review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
              examples:
                Review 3:
                 $ref: '#/components/examples/review3'
                Review 2:
                 $ref: '#/components/examples/review2'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - reviews
      summary: Update the specified review
      description: Update the specified review
      operationId: updateReviewEvent
      parameters:
      - name: eventId
        in: path
        description: Id of the event
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      - name: reviewId
        in: path
        description: Id of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      requestBody:
        description: Payload of the updated review
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        "200":
          description: Review successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
              examples:
                Review 3:
                 $ref: '#/components/examples/review3'
                Review 2:
                 $ref: '#/components/examples/review2'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - reviews
      summary: Deletes a review
      description: Deletes the review with the given id
      operationId: deleteReviewEvent
      parameters:
      - name: eventId
        in: path
        description: Id of the place
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      - name: reviewId
        in: path
        description: Id of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      responses:
        "204":
          description: Review successfully deleted
        "404":
          $ref: '#/components/responses/NotFound'
  /events/{eventId}/place:
    get:
      tags: 
        - events
      summary: Get a place of an event
      description: Get a place of an event given the event id
      operationId: getEventPlace
      parameters:
        - name: eventId
          in: path
          description: Id of the event
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Values retrieved of a particular place
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
              examples:
                Place 1:
                  $ref: '#/components/examples/Estanislao'
                Event2:
                  $ref: '#/components/examples/CatedralSevilla'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags: 
        - events
      summary: Delete a place of an event
      description: Get a place of an event given the event id
      operationId: deleteEventPlace
      parameters:
        - name: eventId
          in: path
          description: Id of the event
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int32
      responses:
        "204":
          $ref: '#/components/responses/ResourceDeleted'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Events:
      type: array
      items:
        $ref: '#/components/schemas/Event'
...
    Place:
      required:
      - address
      - location
      - name
      type: object
      properties:
        name:
          type: string
          description: Name of the place or alias
        address:
          type: string
        category:
          type: string
          nullable: true
          enum:
          - SUPERMARKET
          - PHARMACY
          - ATM
          - BUSINESS
          - BAR
          - RESTAURANT
          - CHURCH
          - CINEMA
          - PARK
          - NIGTH_CLUB
          - MUSEUM
          - UNIVERSITY
          - GYM
          - CONGRESS
          - HOSPITAL
          - POLICE
          - TOURIST_ATTRACTION
          - ACCOMODATION
          - STADIUM
          - GAS_STATION
          - TRAIN_STATION
          - BUS_STATION
          - AIRPORT
          - PARKING
          - TAXI_STAND
          - SUBWAY_STATION
          - UNDEFINED
        email:
          type: string
          description: Contact email of the place
          nullable: true
        website:
          type: string
          description: Website of the place
          nullable: true
        accommodation:
          $ref: '#/components/schemas/Accomodations'
        rating:
          type: number
          format: double
          description: Average rating of a place
          readOnly: true
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        location:
          $ref: '#/components/schemas/Coordinates'
...
  examples:
...
    CatedralSevilla:
      summary: La catedral de sevilla
      description: Example the seville cathedral
      value:
        name: Catedral de Sevilla
        address: "Av. de la Constitución, s/n, 41004 Sevilla"
        website: https://www.catedraldesevilla.es
        location:
          latitude: 37.38586299856296
          longitude: -5.993103877606445
    TestPlace:
      description: Example of Imaginary place
      value:
        name: This is a test of place
        email: test@testing.com
        address: C. Test s/n
        website: https://foobar.com
        location:
          latitude: 25.000777231550654
          longitude: -71.00017170317003
...